gui refactoring
